refactor(testing): add fixtures subpath with shared mock data
