+++ mocks/mock-data/utils/getDate.ts
@@ -1,7 +1,1 @@
-const getDate = (daysOffset = 0, originDate = new Date()) => {
-    const date = new Date(originDate);
-    date.setDate(date.getDate() + daysOffset);
-    return date.toISOString();
-};
-
-export default getDate;
+export { getDate as default } from '@integration-components/testing/fixtures';
