+++ packages/shared/testing/package.json
@@ -8,6 +8,7 @@
     "exports": {
         ".": "./src/index.ts",
         "./package.json": "./package.json",
+        "./fixtures": "./src/fixtures/index.ts",
         "./msw": "./src/msw/index.ts",
         "./playwright": "./src/playwright/index.ts",
         "./storybook-helpers": "./src/storybook-helpers/index.ts",
