+++ packages/shared/testing/src/fixtures/getDate.ts
@@ -1,0 +1,7 @@
+const getDate = (daysOffset = 0, originDate = new Date()) => {
+    const date = new Date(originDate);
+    date.setDate(date.getDate() + daysOffset);
+    return date.toISOString();
+};
+
+export default getDate;
