+++ packages/shared/testing/src/fixtures/index.ts
@@ -1,0 +1,2 @@
+export { BALANCE_ACCOUNTS, BALANCES, BALANCE_ACCOUNTS_SINGLE } from './balanceAccounts';
+export { default as getDate } from './getDate';
