+++ packages/shared/testing/src/index.ts
@@ -1,3 +1,4 @@
+export * as fixtures from './fixtures';
 export * as msw from './msw';
 export * as playwright from './playwright';
 export * as storybookHelpers from './storybook-helpers';
